+++ CHANGELOG.md
@@ -38,6 +38,10 @@ Classify the change according to the following categories:
 ### Fixed
 - Fixed a bug in `cost_curve_constraints.jl` where **m[:PVCapexNoIncentives]** was being added to **m[:InitialCapexNoIncentives]** twice in the PV capex for loop. Updated Multiple PVs test in `runtests.jl` to validate **initial_capital_cost** equals **initial_capital_cost_after_incentives** when all PV incentives are zeroed out.
 
+## chp-follow-thermal-load
+### Added
+- Added attribute `CHP.follow_heating_load` which, when set to `true`, adds new constraints via function `add_chp_heating_load_following_constraints` to enforce a policy in which CHP runs at capacity if the CHP size is less than the heating load, and restricts the `ExistingBoiler` from serving the load when CHP size is greater than the heating load in each time period (this allows for other heating techs like CST to contribute potentially).
+
 ## v0.59.3
 ### Fixed
 - Fixed a bug in which CHP vented heat instead of sending it to the absorption chiller when both the electrical-load-following policy and the absorption-chiller-only policy were enforced.